# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tempo API",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/sets/{set_id}": {
      "patch": {
        "tags": [
          "sets"
        ],
        "summary": "Update Set",
        "operationId": "update_set_api_sets__set_id__patch",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Set Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedSetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sets"
        ],
        "summary": "Delete Set",
        "description": "Soft-delete a set and recompute its exercise's records.\n\nReturns the removed row rather than 204 so the caller has the id to `restore` with, and can\nsee it really was this set. The record it may have held falls back to the next best.",
        "operationId": "delete_set_api_sets__set_id__delete",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prs": {
      "get": {
        "tags": [
          "prs"
        ],
        "summary": "List Prs",
        "operationId": "list_prs_api_prs_get",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "prs"
        ],
        "summary": "Log Pr",
        "description": "Record a PR by hand — the REST twin of the MCP ``log_pr`` tool.\n\nFor records a logged set cannot express: an estimated 1RM, a hold timed outside a session, or\na PR migrated from another app. Overwrites whatever is stored for this exercise + metric.",
        "operationId": "log_pr_api_prs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prs/{pr_id}": {
      "patch": {
        "tags": [
          "prs"
        ],
        "summary": "Update Pr",
        "description": "Correct a hand-entered record in place, without the delete-and-reinsert dance.",
        "operationId": "update_pr_api_prs__pr_id__patch",
        "parameters": [
          {
            "name": "pr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pr Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "prs"
        ],
        "summary": "Delete Pr",
        "description": "Withdraw a hand-entered record; the next best becomes current rather than leaving a gap.",
        "operationId": "delete_pr_api_prs__pr_id__delete",
        "parameters": [
          {
            "name": "pr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pr Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrDeleteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prs/history/{entry_id}": {
      "delete": {
        "tags": [
          "prs"
        ],
        "summary": "Delete Pr History Entry",
        "description": "Remove one chronology entry, then rebuild the record from what is left.",
        "operationId": "delete_pr_history_entry_api_prs_history__entry_id__delete",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrHistoryItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prs/history": {
      "get": {
        "tags": [
          "prs"
        ],
        "summary": "Pr History",
        "operationId": "pr_history_api_prs_history_get",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          },
          {
            "name": "pr_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pr Type"
            }
          },
          {
            "name": "include_deleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Deleted"
            }
          },
          {
            "name": "include_uncounted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Uncounted"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrHistoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoggedSetOut": {
        "properties": {
          "set": {
            "$ref": "#/components/schemas/SetOut"
          },
          "pr": {
            "$ref": "#/components/schemas/PrInfo"
          }
        },
        "type": "object",
        "required": [
          "set",
          "pr"
        ],
        "title": "LoggedSetOut"
      },
...
      "PrCreate": {
        "properties": {
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id"
          },
          "pr_type": {
            "type": "string",
            "title": "Pr Type"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "achieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Achieved At"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "client_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Key"
          }
        },
        "type": "object",
        "required": [
          "exercise_id",
          "pr_type",
          "value",
          "achieved_at"
        ],
        "title": "PrCreate",
        "description": "A hand-entered record. ``unit`` is not a field — it follows from ``pr_type``."
      },
      "PrDeleteOut": {
        "properties": {
          "standing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "standing"
        ],
        "title": "PrDeleteOut",
        "description": "What is standing after a record was withdrawn — ``null`` if nothing supports one."
      },
      "PrHistoryItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "pr_type": {
            "type": "string",
            "title": "Pr Type"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "achieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Achieved At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "set_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "counted": {
            "type": "boolean",
            "title": "Counted"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pr_type",
          "value",
          "unit",
          "achieved_at",
          "source",
          "set_id",
          "session_id",
          "notes",
          "counted",
          "deleted_at"
        ],
        "title": "PrHistoryItem",
        "description": "One entry in a record's chronology, auto or manual.\n\nThis replaces the ``SetOut`` this endpoint used to return. A manual record has no set behind\nit, so a set-shaped payload could not represent one — which is precisely why manual entries\nwere missing from PR history before."
      },
      "PrHistoryOut": {
        "properties": {
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id"
          },
          "pr_type": {
            "type": "string",
            "title": "Pr Type"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PrHistoryItem"
            },
            "type": "array",
            "title": "Items",
            "description": "Oldest first; manual and auto entries interleaved chronologically."
          }
        },
        "type": "object",
        "required": [
          "exercise_id",
          "pr_type",
          "items"
        ],
        "title": "PrHistoryOut"
      },
      "PrInfo": {
        "properties": {
          "is_pr": {
            "type": "boolean",
            "title": "Is Pr"
          },
          "pr_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Type"
          },
          "previous_best": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Best"
          },
          "new_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Value"
          }
        },
        "type": "object",
        "required": [
          "is_pr",
          "pr_type",
          "previous_best",
          "new_value"
        ],
        "title": "PrInfo",
        "description": "The PR verdict for the set that was just logged/updated."
      },
      "PrListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PrOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PrListOut"
      },
      "PrOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id"
          },
          "exercise_name": {
            "type": "string",
            "title": "Exercise Name"
          },
          "exercise_slug": {
            "type": "string",
            "title": "Exercise Slug"
          },
          "pr_type": {
            "type": "string",
            "title": "Pr Type"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "achieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Achieved At"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "illustration_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Illustration Url"
          },
          "illustration_status": {
            "type": "string",
            "title": "Illustration Status"
          },
          "is_custom": {
            "type": "boolean",
            "title": "Is Custom"
          }
        },
        "type": "object",
        "required": [
          "id",
          "exercise_id",
          "exercise_name",
          "exercise_slug",
          "pr_type",
          "value",
          "unit",
          "achieved_at",
          "session_id",
          "notes",
          "source",
          "illustration_url",
          "illustration_status",
          "is_custom"
        ],
        "title": "PrOut"
      },
      "PrUpdate": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "achieved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Achieved At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "clear_notes": {
            "type": "boolean",
            "title": "Clear Notes",
            "default": false
          }
        },
        "type": "object",
        "title": "PrUpdate",
        "description": "Correct a hand-entered record in place. Only supplied fields change."
      },
...
      "SetOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id"
          },
          "set_number": {
            "type": "integer",
            "title": "Set Number"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "reps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reps"
          },
          "hold_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold Seconds"
          },
          "rpe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpe"
          },
          "is_pr": {
            "type": "boolean",
            "title": "Is Pr"
          },
          "pr_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Type"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "is_backfill": {
            "type": "boolean",
            "title": "Is Backfill"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "exercise_id",
          "set_number",
          "weight_kg",
          "reps",
          "hold_seconds",
          "rpe",
          "is_pr",
          "pr_type",
          "notes",
          "is_backfill",
          "created_at"
        ],
        "title": "SetOut"
      },
      "SetUpdate": {
        "properties": {
          "set_number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Number"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "reps": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Reps"
          },
          "hold_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold Seconds"
          },
          "rpe": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpe"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "SetUpdate",
        "description": "Patch a set's metrics; only supplied fields change (PR is recomputed)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}